requirements.txt
SQLAlchemy~=1.4.31
python-dotenv~=0.19.2
h3~=3.7.3